Layouts: introduce Views::Layouts namespace for layout components
Two parts of the application layout move from `Components::*` to
the new `Views::Layouts::*` namespace — a peer to
`Views::Controllers::*` that holds Phlex classes bound to the
layout layer (not to any specific controller, not a reusable
building block):

- `Components::ApplicationSidebar` → `Views::Layouts::ApplicationSidebar`
- `Components::LoginLayout`        → `Views::Layouts::LoginWelcome`

The `LoginLayout` rename is in scope because the class isn't really
a layout — it's a welcome-banner shown to unverified users. The
single render site in `application.html.erb` is currently commented
out; updated to point at the new namespace name for when it's
re-enabled.

Auto-namespacing: no `phlex.rb` initializer change needed.
`Rails.autoloaders.main.push_dir("app/views", namespace: Views)` is
already in place, so Zeitwerk automatically maps
`app/views/layouts/<name>.rb` to `Views::Layouts::<Name>`.

Layout-rendered widgets that AREN'T layout chrome stay in
`app/components/`:

- `Components::ModalConfirm` — global Turbo confirmation modal,
  driven by JS. Lives in the layout but it's UI infrastructure, not
  a layout fragment.
- `Components::ModalProgressSpinner` — same: global JS-driven modal.
- `Components::TranslatorsCredit` — i18n footer widget. Layout-
  adjacent but not layout chrome.

Tests moved + namespaced to `Views::Layouts`. 16 tests pass; lint
clean.
